requirements.txt
torch==2.0.1
dgl>=1.1
numpy
scipy
pandas
pyyaml
networkx
pytest
tqdm